# Conflicts:
#	.gitignore
#	pycaret/anomaly.py
#	pycaret/classification.py
#	pycaret/clustering.py
#	pycaret/internal/preprocess/preprocess.py
#	pycaret/internal/pycaret_experiment/classification_experiment.py
#	pycaret/internal/pycaret_experiment/regression_experiment.py
#	pycaret/internal/pycaret_experiment/supervised_experiment.py
#	pycaret/internal/pycaret_experiment/tabular_experiment.py
#	pycaret/internal/utils.py
#	pycaret/regression.py

+++ pycaret/internal/persistence.py
@@ -2,17 +2,17 @@
 # Author: Moez Ali <moez.ali@queensu.ca> and Antoni Baum (Yard1) <antoni.baum@protonmail.com>
 # License: MIT
 
-from typing import Dict, Optional
-from pycaret.internal.utils import get_logger
-from pycaret.internal.Display import Display
-from sklearn.pipeline import Pipeline
 import gc
+from typing import Dict, Optional
+
+from sklearn.pipeline import Pipeline
+
+from pycaret.internal.utils import get_logger
 
 
 def deploy_model(
-    model, model_name: str, authentication: dict, platform: str = "aws", prep_pipe_=None
+        model, model_name: str, authentication: dict, platform: str = "aws", prep_pipe_=None
 ):
-
     """
     (In Preview)
 
@@ -38,16 +38,16 @@ def deploy_model(
     For GCP users:
     --------------
     Before deploying a model to Google Cloud Platform (GCP), user has to create Project
-    on the platform from consol. To do that, user must have google cloud account or
-    create new one. After creating a service account, down the JSON authetication file
+    on the platform from console. To do that, user must have google cloud account or
+    create new one. After creating a service account, down the JSON authentication file
     and configure  GOOGLE_APPLICATION_CREDENTIALS= <path-to-json> from command line. If
-    using google-colab then authetication can be done using `google.colab` auth method.
+    using google-colab then authentication can be done using `google.colab` auth method.
     Read below link for more details.
 
     https://cloud.google.com/docs/authentication/production
 
     - Google Cloud Project
-    - Service Account Authetication
+    - Service Account Authentication
 
     For Azure users:
     ---------------
@@ -105,8 +105,6 @@ def deploy_model(
     logger.info("Initializing deploy_model()")
     logger.info(f"deploy_model({function_params_str})")
 
-    import sys
-
     # ignore warnings
     import warnings
 
@@ -127,9 +125,7 @@ def deploy_model(
             raise ValueError("Authentication is missing.")
 
     # general dependencies
-    import ipywidgets as ipw
-    import pandas as pd
-    from IPython.display import clear_output, update_display
+    from IPython.display import clear_output
     import os
 
     logger.info("Saving model in active working directory")
@@ -141,30 +137,41 @@ def deploy_model(
 
         logger.info("Platform : AWS S3")
 
-        # checking if awscli available
+        # checking if boto3 is available
         try:
-            import awscli
-        except:
+            import boto3
+        except ModuleNotFoundError:
             logger.error(
-                "awscli library not found. pip install awscli to use deploy_model function."
+                "boto3 library not found. pip install boto3 to use deploy_model function."
             )
             raise ImportError(
-                "awscli library not found. pip install awscli to use deploy_model function."
+                "boto3 library not found. pip install boto3 to use deploy_model function."
             )
 
-        import boto3
-
-        # initiaze s3
+        # initialize s3
         logger.info("Initializing S3 client")
         s3 = boto3.client("s3")
         filename = f"{model_name}.pkl"
         key = f"{model_name}.pkl"
         bucket_name = authentication.get("bucket")
-        s3.upload_file(filename, bucket_name, key)
+
+        if bucket_name is None:
+            logger.error('S3 bucket name missing. Provide `bucket` as part of authentication parameter.')
+            raise ValueError('S3 bucket name missing. Provide `bucket` name as part of authentication parameter.')
+
+        import botocore.exceptions
+        try:
+            s3.upload_file(filename, bucket_name, key)
+        except botocore.exceptions.NoCredentialsError:
+            logger.error('Boto3 credentials not configured. Refer boto3 documentation '
+                         '(https://boto3.amazonaws.com/v1/documentation/api/latest/guide/credentials.html)')
+            logger.error('Model deployment to AWS S3 failed.')
+            raise ValueError('Boto3 credentials not configured. Refer boto3 documentation '
+                             '(https://boto3.amazonaws.com/v1/documentation/api/latest/guide/credentials.html)')
         clear_output()
         os.remove(filename)
-        print("Model Succesfully Deployed on AWS S3")
-        logger.info("Model Succesfully Deployed on AWS S3")
+        print("Model Successfully Deployed on AWS S3")
+        logger.info("Model Successfully Deployed on AWS S3")
         logger.info(str(model))
 
     elif platform == "gcp":
@@ -173,7 +180,8 @@ def deploy_model(
 
         try:
             import google.cloud
-        except:
+
+        except ModuleNotFoundError:
             logger.error(
                 "google-cloud-storage library not found. pip install google-cloud-storage to use deploy_model function with GCP."
             )
@@ -186,21 +194,28 @@ def deploy_model(
         key = f"{model_name}.pkl"
         bucket_name = authentication.get("bucket")
         project_name = authentication.get("project")
+
+        if bucket_name is None or project_name is None:
+            logger.error('Project and Bucket name missing. Provide `bucket` and `project` as part of '
+                         'authentication parameter')
+            raise ValueError('Project and Bucket name missing. Provide `bucket` and `project` as part of '
+                             'authentication parameter')
+
         try:
             _create_bucket_gcp(project_name, bucket_name)
             _upload_blob_gcp(project_name, bucket_name, filename, key)
         except:
             _upload_blob_gcp(project_name, bucket_name, filename, key)
         os.remove(filename)
-        print("Model Succesfully Deployed on GCP")
-        logger.info("Model Succesfully Deployed on GCP")
+        print("Model Successfully Deployed on GCP")
+        logger.info("Model Successfully Deployed on GCP")
         logger.info(str(model))
 
     elif platform == "azure":
 
         try:
             import azure.storage.blob
-        except:
+        except ModuleNotFoundError:
             logger.error(
                 "azure-storage-blob library not found. pip install azure-storage-blob to use deploy_model function with Azure."
             )
@@ -214,6 +229,11 @@ def deploy_model(
         filename = f"{model_name}.pkl"
         key = f"{model_name}.pkl"
         container_name = authentication.get("container")
+
+        if container_name is None:
+            logger.error('Storage Container name missing. Provide `container` as part of authentication parameter')
+            raise ValueError('Storage Container name missing. Provide `container` as part of authentication parameter')
+
         try:
             container_client = _create_container_azure(container_name)
             _upload_blob_azure(container_name, filename, key)
@@ -223,18 +243,17 @@ def deploy_model(
 
         os.remove(filename)
 
-        print("Model Succesfully Deployed on Azure Storage Blob")
-        logger.info("Model Succesfully Deployed on Azure Storage Blob")
+        print("Model Successfully Deployed on Azure Storage Blob")
+        logger.info("Model Successfully Deployed on Azure Storage Blob")
         logger.info(str(model))
 
     logger.info(
-        "deploy_model() succesfully completed......................................"
+        "deploy_model() successfully completed......................................"
     )
     gc.collect()
 
 
 def save_model(model, model_name: str, prep_pipe_=None, verbose: bool = True, **kwargs):
-
     """
     This generic function saves the transformation pipeline and trained model object 
     into the current active directory as a pickle file for later use. 
@@ -294,24 +313,23 @@ def save_model(model, model_name: str, prep_pipe_=None, verbose: bool = True, **
     model_name = f"{model_name}.pkl"
     joblib.dump(model_, model_name, **kwargs)
     if verbose:
-        print("Transformation Pipeline and Model Succesfully Saved")
+        print("Transformation Pipeline and Model Successfully Saved")
 
     logger.info(f"{model_name} saved in current working directory")
     logger.info(str(model_))
     logger.info(
-        "save_model() succesfully completed......................................"
+        "save_model() successfully completed......................................"
     )
     gc.collect()
-    return (model_, model_name)
+    return model_, model_name
 
 
 def load_model(
-    model_name,
-    platform: Optional[str] = None,
-    authentication: Optional[Dict[str, str]] = None,
-    verbose: bool = True,
+        model_name,
+        platform: Optional[str] = None,
+        authentication: Optional[Dict[str, str]] = None,
+        verbose: bool = True,
 ):
-
     """
     This generic function loads a previously saved transformation pipeline and model 
     from the current active directory into the current python environment. 
@@ -360,7 +378,6 @@ def load_model(
     warnings.filterwarnings("ignore")
 
     # exception checking
-    import sys
 
     if platform:
         if not authentication:
@@ -378,9 +395,25 @@ def load_model(
     # cloud providers
     elif platform == "aws":
 
-        import os, boto3
+        import os
+
+        # checking if boto3 is available
+        try:
+            import boto3
+        except ModuleNotFoundError:
+            logger.error(
+                "boto3 library not found. pip install boto3 to use deploy_model function."
+            )
+            raise ImportError(
+                "boto3 library not found. pip install boto3 to use deploy_model function."
+            )
 
         bucketname = authentication.get("bucket")
+
+        if bucketname is None:
+            logger.error('S3 bucket name missing. Provide `bucket` as part of authentication parameter')
+            raise ValueError('S3 bucket name missing. Provide `bucket` name as part of authentication parameter')
+
         filename = f"{model_name}.pkl"
         index = filename.rfind("/")
         s3 = boto3.resource("s3")
@@ -388,7 +421,7 @@ def load_model(
         if index == -1:
             s3.Bucket(bucketname).download_file(filename, filename)
         else:
-            path, key = filename[: index + 1], filename[index + 1 :]
+            path, key = filename[: index + 1], filename[index + 1:]
             if not os.path.exists(path):
                 os.makedirs(path)
             s3.Bucket(bucketname).download_file(key, filename)
@@ -404,6 +437,13 @@ def load_model(
 
         bucket_name = authentication.get("bucket")
         project_name = authentication.get("project")
+
+        if bucket_name is None or project_name is None:
+            logger.error('Project and Bucket name missing. Provide `bucket` and `project` as part of '
+                         'authentication parameter')
+            raise ValueError('Project and Bucket name missing. Provide `bucket` and `project` as part of '
+                             'authentication parameter')
+
         filename = f"{model_name}.pkl"
 
         model_downloaded = _download_blob_gcp(
@@ -419,6 +459,11 @@ def load_model(
     elif platform == "azure":
 
         container_name = authentication.get("container")
+
+        if container_name is None:
+            logger.error('Storage Container name missing. Provide `container` as part of authentication parameter')
+            raise ValueError('Storage Container name missing. Provide `container` as part of authentication parameter')
+
         filename = f"{model_name}.pkl"
 
         model_downloaded = _download_blob_azure(container_name, filename, filename)
@@ -459,7 +504,14 @@ def _create_bucket_gcp(project_name: str, bucket_name: str):
     # bucket_name = "your-new-bucket-name"
     from google.cloud import storage
 
-    storage_client = storage.Client(project_name)
+    import google.auth.exceptions
+    try:
+        storage_client = storage.Client(project_name)
+    except google.auth.exceptions.DefaultCredentialsError:
+        logger.error('Environment variable GOOGLE_APPLICATION_CREDENTIALS not set. For more information,'
+                     ' please see https://cloud.google.com/docs/authentication/getting-started')
+        raise ValueError('Environment variable GOOGLE_APPLICATION_CREDENTIALS not set. For more information,'
+                         ' please see https://cloud.google.com/docs/authentication/getting-started')
 
     buckets = storage_client.list_buckets()
 
@@ -471,12 +523,11 @@ def _create_bucket_gcp(project_name: str, bucket_name: str):
 
 
 def _upload_blob_gcp(
-    project_name: str,
-    bucket_name: str,
-    source_file_name: str,
-    destination_blob_name: str,
+        project_name: str,
+        bucket_name: str,
+        source_file_name: str,
+        destination_blob_name: str,
 ):
-
     """
     Upload blob to GCP storage bucket
 
@@ -511,7 +562,15 @@ def _upload_blob_gcp(
     # destination_blob_name = "storage-object-name"
     from google.cloud import storage
 
-    storage_client = storage.Client(project_name)
+    import google.auth.exceptions
+    try:
+        storage_client = storage.Client(project_name)
+    except google.auth.exceptions.DefaultCredentialsError:
+        logger.error('Environment variable GOOGLE_APPLICATION_CREDENTIALS not set. For more information,'
+                     ' please see https://cloud.google.com/docs/authentication/getting-started')
+        raise ValueError('Environment variable GOOGLE_APPLICATION_CREDENTIALS not set. For more information,'
+                         ' please see https://cloud.google.com/docs/authentication/getting-started')
+
     bucket = storage_client.bucket(bucket_name)
     blob = bucket.blob(destination_blob_name)
 
@@ -523,10 +582,10 @@ def _upload_blob_gcp(
 
 
 def _download_blob_gcp(
-    project_name: str,
-    bucket_name: str,
-    source_blob_name: str,
-    destination_file_name: str,
+        project_name: str,
+        bucket_name: str,
+        source_blob_name: str,
+        destination_file_name: str,
 ):
     """
     Download a blob from GCP storage bucket
@@ -561,8 +620,15 @@ def _download_blob_gcp(
     # source_blob_name = "storage-object-name"
     # destination_file_name = "local/path/to/file"
     from google.cloud import storage
+    import google.auth.exceptions
 
-    storage_client = storage.Client(project_name)
+    try:
+        storage_client = storage.Client(project_name)
+    except google.auth.exceptions.DefaultCredentialsError:
+        logger.error('Environment variable GOOGLE_APPLICATION_CREDENTIALS not set. For more information,'
+                     ' please see https://cloud.google.com/docs/authentication/getting-started')
+        raise ValueError('Environment variable GOOGLE_APPLICATION_CREDENTIALS not set. For more information,'
+                         ' please see https://cloud.google.com/docs/authentication/getting-started')
 
     bucket = storage_client.bucket(bucket_name)
     blob = bucket.blob(source_blob_name)
@@ -599,17 +665,22 @@ def _create_container_azure(container_name: str):
     logger = get_logger()
 
     # Create the container
-    import os, uuid
-    from azure.storage.blob import BlobServiceClient, BlobClient, ContainerClient
+    import os
+    from azure.storage.blob import BlobServiceClient
 
     connect_str = os.getenv("AZURE_STORAGE_CONNECTION_STRING")
+
+    if connect_str is None:
+        logger.error('Environment variable AZURE_STORAGE_CONNECTION_STRING not set')
+        raise ValueError('Environment variable AZURE_STORAGE_CONNECTION_STRING not set')
+
     blob_service_client = BlobServiceClient.from_connection_string(connect_str)
     container_client = blob_service_client.create_container(container_name)
     return container_client
 
 
 def _upload_blob_azure(
-    container_name: str, source_file_name: str, destination_blob_name: str
+        container_name: str, source_file_name: str, destination_blob_name: str
 ):
     """
     Upload blob to Azure storage  container
@@ -634,10 +705,14 @@ def _upload_blob_azure(
 
     logger = get_logger()
 
-    import os, uuid
-    from azure.storage.blob import BlobServiceClient, BlobClient, ContainerClient
+    import os
+    from azure.storage.blob import BlobServiceClient
 
     connect_str = os.getenv("AZURE_STORAGE_CONNECTION_STRING")
+
+    if connect_str is None:
+        logger.error('Environment variable AZURE_STORAGE_CONNECTION_STRING not set')
+        raise ValueError('Environment variable AZURE_STORAGE_CONNECTION_STRING not set')
 
     blob_service_client = BlobServiceClient.from_connection_string(connect_str)
     # Create a blob client using the local file name as the name for the blob
@@ -651,7 +726,7 @@ def _upload_blob_azure(
 
 
 def _download_blob_azure(
-    container_name: str, source_blob_name: str, destination_file_name: str
+        container_name: str, source_blob_name: str, destination_file_name: str
 ):
     """
     Download blob from Azure storage  container
@@ -676,10 +751,15 @@ def _download_blob_azure(
 
     logger = get_logger()
 
-    import os, uuid
-    from azure.storage.blob import BlobServiceClient, BlobClient, ContainerClient
+    import os
+    from azure.storage.blob import BlobServiceClient
 
     connect_str = os.getenv("AZURE_STORAGE_CONNECTION_STRING")
+
+    if connect_str is None:
+        logger.error('Environment variable AZURE_STORAGE_CONNECTION_STRING not set')
+        raise ValueError('Environment variable AZURE_STORAGE_CONNECTION_STRING not set')
+
     blob_service_client = BlobServiceClient.from_connection_string(connect_str)
     # Create a blob client using the local file name as the name for the blob
     blob_client = blob_service_client.get_blob_client(
